requirements.txt
django==2.1.3
djangorestframework==3.9.0
gunicorn==19.9.0
pymysql==0.9.2
